gcs/requirements.txt
liboqs-python>=0.7.2
cryptography
pycryptodome
